refactor: rename onehot.py to binary.py
more accurate naming, since it's a bitvector rather than a onehot rep
